test dotnet grpc stackup grpc done file removed
